Refactoring models.py into multiple files
