+++ spotseeker_server/models/__init__.py
@@ -1,0 +1,5 @@
+from .auth import TrustedOAuthClient
+from .item import Item, ItemExtendedInfo, ItemImage
+from .space import SharedSpace, SharedSpaceRecipient, SpaceReview
+from .spot import FavoriteSpot, Spot, SpotAvailableHours, SpotExtendedInfo, \
+    SpotImage, SpotType
